requirements.txt
Flask
google-generativeai
python-dotenv
Flask-Cors
Pillow
